requirements.txt
openai==1.70.0
Pillow==11.1.0
